backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets>=13.0
httpx>=0.28.1
google-genai
chromadb==0.6.3
sentence-transformers==3.0.0
openai-whisper
sounddevice==0.4.7
numpy==1.26.4
sqlalchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1
scipy